requirements.txt

# Requirements for the project
pandas>=1.3.0
openpyxl>=3.0.0
PyMuPDF>=1.18.0
python-docx>=0.8.11
spacy>=3.0.0
streamlit>=0.80.0

pyttsx3